XPAY-2627 feat: sdk structure updates
